+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Change Log
 
+## [NEXT]
+- Save entity data in query history
+
 ## [2022.2.14]
 - Fix situation where Pathom 3 renders a Trace view with blank data
 - Support report from strict errors in Pathom 3